Revert "Merge pull request #35 from MossesAryo/kaprog"
This reverts commit 388035af01b40d3bfbefc57c8b09b11213c46b7f, reversing
changes made to 946e1e6a0b0529d64b5f180ad6d1e4baa0feccb5.

+++ database/migrations/2025_07_29_022231_walikelas.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up() {
         Schema::create('walikelas', function (Blueprint $table) {
-            $table->bigInteger('nip_walikelas')->unique();
+            $table->integer('nip_walikelas')->unique();
             $table->string('username');
             $table->string('id_kelas');
             $table->string('nama_walikelas');
