+++ database/migrations/2025_07_29_022301_guru_bk.php
@@ -4,14 +4,15 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-
+// 6. Guru BK
 return new class extends Migration {
     public function up() {
         Schema::create('guru_bk', function (Blueprint $table) {
-            $table->bigInteger('nip_bk')->unique();
+            $table->integer('nip_bk')->unique();
             $table->string('username');
             $table->string('nama_guru_bk');
             $table->timestamps();
+
             $table->foreign('username')->references('username')->on('users')->onDelete('cascade')->onUpdate('cascade');
         });
     }
